split kimchi_params file
